add `__.record: (a -> b) -> [a] -> {b | a}`
Records the graph of a callable given an array of input values.
Essentially implements the functor `(a -> b) -> {b | a'}` whenever `a'`
is a finite subtype of the union type `a`, coercible to string keys.

N.B. assumes input values are admissible keys. Otherwise resort to
standard `map`, or provide a indexing function as second argument:
```
__.record: (a -> b) -> (a -> k) -> [a] -> {b | k}

__.record = (f, g) => xs => { g(x) : f(x) } for x in xs
```